+++ src/App.js
@@ -1,6 +1,5 @@
 import React, { useEffect } from "react";
 import "./App.css";
-
 import { Switch, Route } from "react-router-dom";
 import Navigation from "./components/Navigation";
 import Loading from "./components/Loading";
@@ -12,14 +11,9 @@ import BecomeSitter from "./pages/BecomeSitter";
 import HomePage from "./pages/HomePage";
 import OurServices from "./components/OurServices";
 import Dashboard from "./pages/Dashboard";
-import Phone from "./components/Phone";
-import Service from "./components/Services";
 import UserById from "./pages/UserById";
 import PetOwner from "./pages/PetOwner";
-import Adderss from "./components/Address";
-import Pets from "./components/Pets";
-import Contact from "./pages/Contact"
-
+import Contact from "./pages/Contact";
 import { useDispatch, useSelector } from "react-redux";
 import { selectAppLoading } from "./store/appState/selectors";
 import { getUserWithStoredToken } from "./store/user/actions";
@@ -39,10 +33,6 @@ function App() {
       {isLoading ? <Loading /> : null}
       <Switch>
         <Route exact path="/" component={HomePage} />
-        <Route path="/address" component={Adderss} />
-        <Route path="/phone" component={Phone} />
-        <Route path="/services" component={Service} />
-        <Route path="/pets" component={Pets}/>
         <Route path="/become_a_sitter" component={BecomeSitter} />
         <Route path="/petowner" component={PetOwner} />
         <Route path="/our_services" component={OurServices} />
@@ -51,7 +41,7 @@ function App() {
         <Route path="/dashboard/:id" component={Dashboard} />
         <Route path="/signup" component={SignUp} />
         <Route path="/login" component={Login} />
-        <Route path="/contact/:id" component={Contact}/>
+        <Route path="/contact/:id" component={Contact} />
       </Switch>
     </div>
   );
