--- src/components/Address/address.css
@@ -1,13 +1,0 @@
-.form-container {
-  display: flex;
-  justify-content: center;
-}
-.form-container-2 {
-  max-width: 500px;
-}
-.sign-up {
-  font-size: 1.75rem;
-  font-weight: 500;
-  margin-bottom: 15px;
-  font-family: "Prata", serif;
-}
